+++ metadata.rb
@@ -1,7 +1,7 @@
 name             'godns'
 maintainer       'Leo Unbekandt'
-maintainer_email 'leo.unbekandt@appsdeck.eu'
+maintainer_email 'leo@scalingo.com'
 license          'MIT'
 description      'Install GoDNS'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.1.9'
+version          '1.2.0'